requirements.test.txt
# From our manifest.json for our custom component
ccxt=2.4.47

# Strictly for tests
codecov==2.1.12
coverage==6.3.2
pytest==7.1.1
pytest-cov==2.12.1
pytest-homeassistant-custom-component==0.8.6
